save&insert tests, deleted tests with databaseTester interface
